\name{bars}
\alias{bars}
\title{Bayesian Adaptive Regression Splines}
\description{Solves the generalized nonparametric regression (curve-fitting)
problem by assuming the function f(x) may be approximated with a cubic
spline.  Utilizes reversible-jump MCMC on the sets of knots, allowing the
spline to adapt to the data across the domain of x.
}
\usage{
bars<-(x,y,iknots=25,prior="uniform",
priorparam=c(1,60),burnin=200,sims=2000,tau=50.0,c=0.4,fits=TRUE,
peak=FALSE,conf=0.95,gridfit=FALSE,ngridpoints=150)
}
\arguments{
Required: x,y

Optional:\cr
          Initial knots settings: iknots\cr
	  Settings on prior for knots: prior, priorparam\cr
	  MCMC settings: burnin, nsims, tau, c\cr
	  Output settings: fits, peak, conf\cr
	  Other: gridfit, ngridpoints\cr

  \item{x}{ a vector of the independent variable, in increasing order }
  \item{y}{ a vector of the dependent variable }
  \item{iknots}{ the initial number of knots for the spline [default =
25] }
  \item{prior}{ the type of prior being used for the knots; the only
acceptable answers are "poisson", "uniform", and "user" [default is
"uniform"] }
  \item{priorparam}{ the parameter for the prior  (See Details) }
  \item{burnin}{ the desired length of the burn-in for the MCMC chain
[default = 200] }
  \item{nsims}{ the number of simulations (number of draws from the
posterior) desired for the MCMC chain
[default = 2000] }
  \item{tau}{ a parameter that controls the spread for the knot
proposal distribution [default = 50.0] }
  \item{c}{a paramater that controls the probability of birth and
death candidates [default = 0.4] }
  \item{fits}{ a logical.  The default 'TRUE' will cause the program
to return the fitted values for each data point for each run of the
simulation.  Please note that if the number of data points and/or
simulations is large, there may be a lengthy delay as the necessary
data are read.}
  \item{peak}{ a logical.  'TRUE' will cause the program to
return the location and height of the highest point on the fitted
curve [default = 'FALSE'] }
  \item{conf}{ for use with peak.  Sets the probability for the
credible intervals for the location and height of the peak  [default =
0.95 for 95\% credible intervals] }
  \item{ngridpoints}{ the number of bins one desires the x-axis to be divided
into.  Used to handle unbinned data and calculate posterior modes
[default = 150] }
}
\details{Only x and y are required arguments.  Unspecified arguments
will use default values:

     bars(x, y)\cr
     bars(x, y, iknots=5)

For the parameter for the prior:

     a. if using Poisson, the choice for lambda = mean 

     b. if using Uniform, a vector of length 2 which includes the
	minimum number of knots followed by the maximum number of
	knots [default = c(1,60)]

     c. if using user-defined prior, a matrix with 2 columns.  The
	first column should be the number of knots and the second column
	should be the probability of obtaining this number of knots.
	Note the following example:

	\tabular{lr}{
		 2		\tab	0.05 \cr
		 3		\tab	0.15 \cr
		 4		\tab	0.30 \cr
		 5		\tab	0.30 \cr
		 6		\tab	0.10 \cr
		 7		\tab	0.10 \cr
	}

Note that due to the large amount of output generated by the program,
it is desirable to save the results of the program into a variable, such as:

out <- bars(x,y,.....)
}
\value{
  a list containing the following:
  Always:  postmeans, postmodes, simnum, nknots, sampknots,
sampBICs, sampllikes
  Optional, if fits = TRUE: sampfits
  Optional, if gridfit = TRUE: gridmeans, gridmodes
  Optional, if peak = TRUE:  samplpeaks, samphpeaks,
peaklocationquantile, peaklocationmean, peaklocationmode,
peakheightquantile, peakheightmean, peakheightmode

  \item{postmeans}{ vector of the posterior means of f(x) evaluated at the x
values }
  \item{postmodes}{ vector of the posterior modes evaluated at the x
values }
  \item{simnum}{  vector of each simulation number (the number of that
draw from the posterior), beginning at burnin + 1 and ending at burnin
+ nsims }
  \item{nknots}{ vector of the number of knots used at each draw from
the posterior (does not include burnin iterations) }
  \item{sampknots}{ matrix containing the position of the knots at
each iteration. Length of the matrix is equal to the number of
iterations, not including burnin iterations, with the width of the
matrix equal to the maximum number of knots at any iteration.  NAs are
used to fill in the matrix at iteration numbers that have less than
the maximum number of knots. }
  \item{sampBICs}{  vector of the calculated BIC at each draw from the
posterior (does not include burnin iterations) }
  \item{sampllikes}{ vector of the calculated loglikelihood at each
draw from the posterior (does not include burnin iterations) }
  \item{sampfits}{ optional, if fits='TRUE':  matrix of fits for each
draw from the posterior (does not include burnin iterations), with the
rows of the matrix corresponding to the individual draw from the
posterior.  The columns represent the fits at each value of x.}
  \item{gridmeans}{ optional, if gridfit='TRUE':  vector of the
posterior means evaluated at equally spaced intervals with number of
intervals determined by value of "ngridpoints" }
  \item{gridmodes}{ optional, if gridfit='TRUE': vector of the
posterior modes evaluated at equally spaced intervals with number of
intervals determined by value of "ngridpoints" }
  \item{samplpeaks}{ optional, if peaks='TRUE': vector of the x location
of the highest point in the fitted curve for each draw from the
posterior (does not include burnin iterations)}
  \item{samphpeaks}{ optional, if peaks='TRUE': vector of the y value
(height) of the highest point in the fitted curve for each draw from
the posterior (does not include burnin iterations) }
  \item{peaklocationquantile}{ optional, if peaks='TRUE': a credible
interval for the x location of the highest peak; width of the interval
is dependent upon the setting chosen for conf }
  \item{peaklocationmean}{ optional, if peaks='TRUE': the mean x location
for the highest peak }
  \item{peaklocationmode}{ optional, if peaks='TRUE': the mode x location
for the highest peak }
  \item{peakheightquantile}{ optional, if peaks='TRUE': a credible
interval for the y value (height) of the heighest peak; width of the
interval is dependent upon the setting chosen for conf }
  \item{peakheightmean}{ optional, if peaks='TRUE': the mean y value
(height) of the heighest peak }
  \item{peakheightmode}{ optional, if peaks='TRUE': the mode y value
(height) of the highest peak }

}
\references{ 
This function is designed to work with Wallstrom and Kass's Barslib, (C) 2003.

Barslib uses Hansen and Kooperberg's Logspline, (C) 1997, as the
default method for selecting initial knots.

Barslib uses Bates and Venables Routines for manipulating B-splines,
(C) 1998. The routines were released under the GNU General Public
License version 

see \url{ http://www.stat.cmu.edu/~kass/bars/bars.html }
}
\author{ Garrick Wallstrom }
\examples{
x<-c(1:110)
mu<-20*abs(exp(-x/100)*cos(x/5))
y<-rnorm(110,mu,1)
out = bars(x,y,prior="uniform",priorparam=c(1,15),fits=TRUE)

upper=apply(out$sampfits,2,quantile,0.975)
lower=apply(out$sampfits,2,quantile,0.025)

matplot(x,cbind(upper,out$postmeans,lower),lty=c(2,1,2),type='l')
points(x,y)
}
\keyword{ regression }